Don't invalidate existing textures with partial texture data
The current code just assumes all texture deltas are whole images,
which is not true, and writing partial texture data can overwrite
the existing textures with garbage data.

This change makes the texture update loop ignore non-zero offset deltas
which are assumed to be partial data, since they are at a non-zero offset.

Proper partial updates can be implemented in the future.